Implement `RunListItem` component and `RunDataUi` model for run overview
- Create `RunListItem` Composable to display run details, including a map image, duration, date, and a data grid of statistics (distance, pace, speed, elevation).
- Implement `RunDataUi` and `RunUi` (implied via mapper) models to represent run data in the presentation layer.
- Add support for a long-press context menu to delete a run entry.
- Update `Run` domain model to change `distanceInMeters` type from `Double` to `Int`.
- Add required string resources for run details, including labels for average speed, max speed, and error states.
- Integrate `SubcomposeAsyncImage` for asynchronous map loading with progress and error indicators.

+++ core/domain/src/main/java/dev/saketanand/core/domain/run/Run.kt
@@ -9,7 +9,7 @@ data class Run(
     val id: String?,
     val duration: Duration,
     val dateTimeUtc: ZonedDateTime,
-    val distanceInMeters: Double,
+    val distanceInMeters: Int,
     val location: Location,
     val maxSpeedKmh: Double,
     val totalElevationMeters: Int,
